+++ src/app/(auth)/app/(sidebar)/attendances/live-map/_components/live-map.tsx
@@ -319,8 +319,10 @@ export function LiveMap({
         style={{ height: '100%', width: '100%' }}
       >
         <TileLayer
-          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+          attribution="&copy; Google Maps"
+          url="https://mt1.google.com/vt/lyrs=m&x={x}&y={y}&z={z}"
+          maxZoom={20}
+          subdomains={['mt0', 'mt1', 'mt2', 'mt3']}
         />
 
         {clusters.map((cluster) => {